Let linky listen for its environment/mode on both LINKY_ENV and NODE_ENV
Summary:
**Problem:**
Currently, when `npm run setup:links` is hooked up as a `postinstall` script, it always runs in the same mode as the install command, with no way to instruct it otherwise. If `NODE_ENV=development npm ci`, then the `linky` postinstall also runs in `development`. However, there are cases where it would be useful for this not to happen -- where `devDependencies` should be installed, but `devLinks` should not -- or vice versa.

**Solution**
This adds the ability to set linky's mode with the env var `LINKY_ENV`. `LINKY_ENV` is given top priority and will supersede any competing instruction in `NODE_ENV` from linky's perspective. If `LINKY_ENV` is not specified, then `NODE_ENV` will be used as usual. This adds the granularity needed to control linky separately from the install command it may be hooked into via `postinstall`

Example of doing so:
```
LINKY_ENV=production NODE_ENV=development npm ci
```
This installs all dependencies, but only production links.

Also, add a simple test for this

Test Plan: `npm t`

Reviewers: #product_eng, kimcoop

Reviewed By: #product_eng, kimcoop

Subscribers: kimcoop

Differential Revision: https://ikoshq.phacility.com/D2150